- Migrated to Java 17 - Migrated to Spring Boot 3.1.0 - Serenity Cucumber setup added - Changed package name
