+++ src/main/java/nowire/space/learnromanian/LearnromanianRestApplication.java
@@ -1,0 +1,11 @@
+package nowire.space.learnromanian;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LearnromanianRestApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(LearnromanianRestApplication.class, args);
+	}
+}
