+++ src/main/java/nowire/space/learnromanian/model/Role.java
@@ -1,0 +1,41 @@
+package nowire.space.learnromanian.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+
+@Data
+@NoArgsConstructor
+@Entity
+@Table(name="role")
+public class Role implements GrantedAuthority {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="role_id")
+    protected Integer roleId;
+
+    @Column(name="role_name")
+    protected String roleName;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Role )) return false;
+        return roleId != null && roleId.equals(((Role) o).getRoleId());
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
+
+    @Override
+    public String getAuthority() {
+        return roleName;
+    }
+}
