+++ src/main/java/nowire/space/learnromanian/repository/RoleRepository.java
@@ -1,0 +1,6 @@
+package nowire.space.learnromanian.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import nowire.space.learnromanian.model.Role;
+
+public interface RoleRepository extends JpaRepository<Role, Integer> { }
