+++ src/main/java/nowire/space/learnromanian/request/UserAuthCheckRequest.java
@@ -1,0 +1,20 @@
+package nowire.space.learnromanian.request;
+
+public class UserAuthCheckRequest {
+    protected String userEmail;
+
+    public UserAuthCheckRequest() {
+    }
+
+    public UserAuthCheckRequest(String userEmail) {
+        this.userEmail = userEmail;
+    }
+
+    public String getUserEmail() {
+        return userEmail;
+    }
+
+    public void setUserEmail(String userEmail) {
+        this.userEmail = userEmail;
+    }
+}
