+++ src/main/java/nowire/space/learnromanian/response/MessageResponse.java
@@ -1,0 +1,15 @@
+package nowire.space.learnromanian.response;
+
+public class MessageResponse {
+    private String message;
+
+    public MessageResponse(String message) {}
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
