--- src/main/java/ro/ugal/learnromanian/model/Role.java
@@ -1,42 +1,0 @@
-package ro.ugal.learnromanian.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-
-import javax.persistence.*;
-
-@Data
-@NoArgsConstructor
-@Entity
-@Table(name="role")
-public class Role implements GrantedAuthority {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="role_id")
-    protected Integer roleId;
-
-    @Column(name="role_name")
-    protected String roleName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Role )) return false;
-        return roleId != null && roleId.equals(((Role) o).getRoleId());
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-
-    @Override
-    public String getAuthority() {
-        return roleName;
-    }
-}
