--- src/main/java/ro/ugal/learnromanian/model/User.java
@@ -1,93 +1,0 @@
-package ro.ugal.learnromanian.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.persistence.*;
-import java.util.Collection;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@Entity
-@Table(name="user")
-//TO DO password should not be returned
-public class User implements UserDetails {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="user_id")
-    protected Integer userId;
-
-    @Column(name="user_last_name")
-    protected String userLastName;
-
-    @Column(name="user_family_name")
-    protected String userFamilyName;
-
-    @JsonIgnore
-    @Column(name="user_password")
-    protected String userPassword;
-
-    @Column(name="user_phone_number")
-    protected String userPhoneNumber;
-
-    @Column(name="user_email", unique = true)
-    protected String userEmail;
-
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
-    protected List<Role> roles;
-
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "user_photo_id")
-    protected UserPhoto photo;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles;
-    }
-
-    public void addRole(Role role) {
-        roles.add(role);
-        role.setUser(this);
-    }
-
-    public void removeRole(Role role) {
-        roles.remove(role);
-        role.setUser(null);
-    }
-
-    @Override
-    public String getPassword() {
-        return userPassword;
-    }
-
-    @Override
-    public String getUsername() {
-        return userEmail;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
-
