--- src/main/java/ro/ugal/learnromanian/response/MessageResponse.java
@@ -1,15 +1,0 @@
-package ro.ugal.learnromanian.response;
-
-public class MessageResponse {
-    private String message;
-
-    public MessageResponse(String message) {}
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
