--- src/main/java/ro/ugal/learnromanian/util/Enum.java
@@ -1,17 +1,0 @@
-package ro.ugal.learnromanian.util;
-
-public class Enum {
-    public class Role{
-        public static final String ADMIN = "ADMIN";
-        public static final String MODERATOR = "MODERATOR";
-        public static final String PROFESSOR = "PROFESSOR";
-        public static final String STUDENT = "STUDENT";
-    }
-
-    public class DBRole{
-        public static final String ADMIN = "ROLE_ADMIN";
-        public static final String MODERATOR = "ROLE_MODERATOR";
-        public static final String PROFESSOR = "ROLE_PROFESSOR";
-        public static final String STUDENT = "ROLE_STUDENT";
-    }
-}
