--- src/main/java/ro/ugal/learnromanian/util/Message.java
@@ -1,13 +1,0 @@
-package ro.ugal.learnromanian.util;
-
-public class Message {
-    public static String USER_AUTHENTICATION_TRUE = "User is authenticated.";
-    public static String USER_AUTHENTICATION_FALSE = "User is NOT authenticated.";
-
-    public static String USER_LOGIN_TRUE = "User was logged in.";
-    public static String USER_LOGIN_FALSE = "User was NOT found.";
-
-    public static String USER_REGISTRATION_TRUE = "User was successfully registered.";
-    public static String USER_REGISTRATION_FALSE = "Email address is already registered.";
-    public static String USER_REGISTRATION_ERROR = "An error occurred.";
-}
