+++ src/test/java/nowire/space/learnromanian/CucumberTestSuite.java
@@ -1,0 +1,13 @@
+package nowire.space.learnromanian;
+
+import io.cucumber.junit.CucumberOptions;
+import net.serenitybdd.cucumber.CucumberWithSerenity;
+import org.junit.runner.RunWith;
+
+@RunWith(CucumberWithSerenity.class)
+@CucumberOptions(
+        plugin = {"pretty"},
+        features = "src/test/resources/features"
+)
+public class CucumberTestSuite {
+}
