--- src/test/java/ro/ugal/learnromanian/LearnromanianRestApplicationTests.java
@@ -1,13 +1,0 @@
-package ro.ugal.learnromanian;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LearnromanianRestApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
