+++ src/test/resources/cucumber.properties
@@ -1,0 +1,1 @@
+cucumber.publish.quiet=true
